requirements.txt
numpy
opencv-python
scikit-learn
pillow
matplotlib
tqdm
wandb
torchviz
nvitop
ipykernel

# Session 1
transformers
huggingface_hub
timm
ultralytics
pycocotools
albumentations
thop # For calculating FLOPs